# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Jobify API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/applications": {
      "get": {
        "tags": [
          "applications"
        ],
        "summary": "List Applications",
        "operationId": "list_applications_api_v1_applications_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ApplicationRead"
                  },
                  "type": "array",
                  "title": "Response List Applications Api V1 Applications Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "applications"
        ],
        "summary": "Create Application",
        "description": "Track a job (save it, or record that you applied). Sends nothing to the employer.",
        "operationId": "create_application_api_v1_applications_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplicationCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/applications/{application_id}": {
      "get": {
        "tags": [
          "applications"
        ],
        "summary": "Read Application",
        "operationId": "read_application_api_v1_applications__application_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "application_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Application Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "applications"
        ],
        "summary": "Update Application",
        "operationId": "update_application_api_v1_applications__application_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "application_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Application Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplicationUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "applications"
        ],
        "summary": "Delete Application",
        "operationId": "delete_application_api_v1_applications__application_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "application_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Application Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/applications/{application_id}/reminder": {
      "put": {
        "tags": [
          "applications"
        ],
        "summary": "Set Reminder",
        "operationId": "set_reminder_api_v1_applications__application_id__reminder_put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "application_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Application Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Reminder"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "applications"
        ],
        "summary": "Clear Reminder",
        "operationId": "clear_reminder_api_v1_applications__application_id__reminder_delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "application_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Application Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/agent/tasks": {
      "get": {
        "tags": [
          "agent"
        ],
        "summary": "List Tasks",
        "operationId": "list_tasks_api_v1_agent_tasks_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TaskRead"
                  },
                  "type": "array",
                  "title": "Response List Tasks Api V1 Agent Tasks Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "agent"
        ],
        "summary": "Create Task",
        "description": "Start a task. It runs in the background: poll GET /agent/tasks/{id} for progress.",
        "operationId": "create_task_api_v1_agent_tasks_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/agent/tasks/{task_id}": {
      "get": {
        "tags": [
          "agent"
        ],
        "summary": "Get Task",
        "operationId": "get_task_api_v1_agent_tasks__task_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/approvals": {
      "get": {
        "tags": [
          "approvals"
        ],
        "summary": "List Pending",
        "operationId": "list_pending_api_v1_approvals_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ApprovalRead"
                  },
                  "type": "array",
                  "title": "Response List Pending Api V1 Approvals Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/approvals/{approval_id}/approve": {
      "post": {
        "tags": [
          "approvals"
        ],
        "summary": "Approve",
        "description": "The backend performs the action only after this call. The AI cannot approve itself.",
        "operationId": "approve_api_v1_approvals__approval_id__approve_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "approval_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Approval Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/approvals/{approval_id}/reject": {
      "post": {
        "tags": [
          "approvals"
        ],
        "summary": "Reject",
        "operationId": "reject_api_v1_approvals__approval_id__reject_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "approval_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Approval Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ActionRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "step": {
            "type": "integer",
            "title": "Step"
          },
          "tool": {
            "type": "string",
            "title": "Tool"
          },
          "permission": {
            "type": "string",
            "title": "Permission"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          }
        },
        "type": "object",
        "required": [
          "id",
          "step",
          "tool",
          "permission",
          "status",
          "result"
        ],
        "title": "ActionRead"
      },
      "ApplicationCreate": {
        "properties": {
          "jobId": {
            "type": "string",
            "format": "uuid",
            "title": "Jobid"
          },
          "status": {
            "$ref": "#/components/schemas/ApplicationStatus",
            "default": "saved"
          }
        },
        "type": "object",
        "required": [
          "jobId"
        ],
        "title": "ApplicationCreate"
      },
      "ApplicationRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "jobId": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Jobid"
          },
          "jobTitle": {
            "type": "string",
            "title": "Jobtitle"
          },
          "company": {
            "type": "string",
            "title": "Company"
          },
          "location": {
            "type": "string",
            "title": "Location"
          },
          "status": {
            "$ref": "#/components/schemas/ApplicationStatus"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Updatedat"
          },
          "appliedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Appliedat"
          },
          "notes": {
            "type": "string",
            "title": "Notes"
          },
          "reminder": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Reminder"
              },
              {
                "type": "null"
              }
            ]
          },
          "timeline": {
            "items": {
              "$ref": "#/components/schemas/TimelineEvent"
            },
            "type": "array",
            "title": "Timeline"
          },
          "matchScore": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Matchscore"
          },
          "matchLevel": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MatchLevel"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "jobId",
          "jobTitle",
          "company",
          "location",
          "status",
          "createdAt",
          "updatedAt",
          "appliedAt",
          "notes",
          "reminder",
          "timeline",
          "matchScore",
          "matchLevel"
        ],
        "title": "ApplicationRead"
      },
      "ApplicationStatus": {
        "type": "string",
        "enum": [
          "saved",
          "preparing",
          "applied",
          "interview",
          "offer",
          "rejected"
        ],
        "title": "ApplicationStatus"
      },
      "ApplicationUpdate": {
        "properties": {
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApplicationStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "notes": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 10000
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "title": "ApplicationUpdate"
      },
      "ApprovalRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "task_id": {
            "type": "string",
            "format": "uuid",
            "title": "Task Id"
          },
          "tool": {
            "type": "string",
            "title": "Tool"
          },
          "permission": {
            "type": "string",
            "title": "Permission"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "preview": {
            "additionalProperties": true,
            "type": "object",
            "title": "Preview"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "task_id",
          "tool",
          "permission",
          "status",
          "preview",
          "expires_at"
        ],
        "title": "ApprovalRead",
        "description": "What the app shows on the approval screen: the action, who or what it affects, a preview."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MatchLevel": {
        "type": "string",
        "enum": [
          "strong",
          "good",
          "partial",
          "weak"
        ],
        "title": "MatchLevel"
      },
...
      "Reminder": {
        "properties": {
          "dueAt": {
            "type": "string",
            "format": "date-time",
            "title": "Dueat"
          },
          "note": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "type": "object",
        "required": [
          "dueAt"
        ],
        "title": "Reminder"
      },
...
      "TaskCreate": {
        "properties": {
          "goal": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 3,
            "title": "Goal"
          }
        },
        "type": "object",
        "required": [
          "goal"
        ],
        "title": "TaskCreate"
      },
      "TaskRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "goal": {
            "type": "string",
            "title": "Goal"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "result": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "steps_used": {
            "type": "integer",
            "title": "Steps Used"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "actions": {
            "items": {
              "$ref": "#/components/schemas/ActionRead"
            },
            "type": "array",
            "title": "Actions",
            "default": []
          },
          "pending_approval": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApprovalRead"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "goal",
          "status",
          "result",
          "error",
          "steps_used",
          "created_at"
        ],
        "title": "TaskRead"
      },
      "TimelineEvent": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "at": {
            "type": "string",
            "format": "date-time",
            "title": "At"
          },
          "kind": {
            "$ref": "#/components/schemas/TimelineKind"
          },
          "text": {
            "type": "string",
            "title": "Text"
          }
        },
        "type": "object",
        "required": [
          "id",
          "at",
          "kind",
          "text"
        ],
        "title": "TimelineEvent"
      },
      "TimelineKind": {
        "type": "string",
        "enum": [
          "saved",
          "status",
          "note",
          "reminder"
        ],
        "title": "TimelineKind"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}